fix Duration type mismatch bug in
enable DStreamTextFileSample and add Window API

change duration type to seconds for streaming bacth interval, Window, Checkpoint and Remember APIs

fix Duration type mismatch bug in

enable DStreamTextFileSample and add Window API

Update JvmBridgeUtils.cs

use GetJavaDuration utility method

use GetJavaDuration utility method

include Window API in unit test

revert DStreamTextFileSample to experimental

Window API included in unit test

+++ csharp/AdapterTest/EventHubsUtilsTest.cs
@@ -26,7 +26,7 @@ namespace AdapterTest
                                 .Returns(mockDstreamProxy);
 
             var mockSparkClrProxy = new Mock<ISparkCLRProxy>();
-            mockSparkClrProxy.Setup(m => m.CreateStreamingContext(It.IsAny<SparkContext>(), It.IsAny<long>()))
+            mockSparkClrProxy.Setup(m => m.CreateStreamingContext(It.IsAny<SparkContext>(), It.IsAny<int>()))
                 .Returns(streamingContextProxy.Object);
             SparkCLREnvironment.SparkCLRProxy = mockSparkClrProxy.Object;
